ether: lights: Ensure LED is really off when stopping blinking
If you turn off brightness / blinking mode then the kernel driver
will leave the LED in its current state and stop changing it.  That
results in the LED potentially staying lid when we have intended
to disable it.

Instead, tell is that we want our blinking to have a duty cycle of
0 to ensure that it is off before we stop blinking it.

BUGBASH-787

Change-Id: If931614bb1312805456fa148cec1ed7e72c7a64f
Signed-off-by: Joe Maples <joe@frap129.org>